requirements.txt
pip-system-certs>=4.0
aiogram==3.0.0rc2
pydantic-settings>=2.0.3
windows-curses>=2.3.1; platform_system == 'Windows'
Google-Images-Search==1.4.6
salesforce-lavis>=1.0.2
httpx>=0.24.1
orjson>=3.9.5